requirements.txt
numpy==1.26.4
PyQt5==5.15.11
mysql-connector-python==9.5.0
python-dotenv==1.0.0